Add NetlinkFramed::new_multicast.
This adds a new new_multicast constructor to the NetlinkFramed type
which allows for a specific multicast groups value to be passed to the
created SocketAddr instance.

Additionally updated NetlinkFramed::new to leverage
NetlinkFramed::new_multicast with a multicast_groups value of 0 in order
to keep the code DRY and persist the existing functionality.